requirements.txt
Flask==2.3.4
pandas==2.0.3
Werkzeug==2.3.4
python-docx==0.8.11
PyMuPDF==1.21.0
requests==2.28.2
python-dotenv==1.0.0
